git-svn-id: svn://svn.code.sf.net/p/svxlink/svn/branches/improved_voter@1715 f8843c6b-282e-0410-804b-c0d5895947d2

+++ src/svxlink/trx/SigLevDet.h
@@ -140,7 +140,7 @@ class SigLevDet : public SigC::Object, public Async::AudioSink
      * @brief 	Read the latest measured signal level
      * @return	Returns the latest measured signal level
      */
-    virtual double lastSiglev(void) const = 0;
+    virtual float lastSiglev(void) const = 0;
     
     /**
      * @brief   Reset the signal level detector
